+++ src/onegov.election_day/onegov/election_day/views/homepage.py
@@ -8,9 +8,22 @@ from onegov.election_day.models import Principal, Archive
                      permission=Public)
 def view_principal(self, request):
 
+    archive = Archive(request.app.session())
+    latest = archive.latest(group=False)
+    last_modified = archive.last_modified(latest)
+    latest = archive.group_items(latest, reverse=True)
+
+    @request.after
+    def add_last_modified(response):
+        if last_modified:
+            response.headers.add(
+                'Last-Modified',
+                last_modified.strftime("%a, %d %b %Y %H:%M:%S GMT")
+            )
+
     return {
         'layout': DefaultLayout(self, request),
-        'archive_items': Archive(request.app.session()).latest(),
+        'archive_items': latest,
         'show_base_link': True
     }
 
@@ -18,10 +31,23 @@ def view_principal(self, request):
 @ElectionDayApp.json(model=Principal, permission=Public, name='json')
 def view_principal_json(self, request):
 
+    archive = Archive(request.app.session())
+    latest = archive.latest(group=False)
+    last_modified = archive.last_modified(latest)
+    latest = archive.to_json(latest, request)
+
+    @request.after
+    def add_last_modified(response):
+        if last_modified:
+            response.headers.add(
+                'Last-Modified',
+                last_modified.strftime("%a, %d %b %Y %H:%M:%S GMT")
+            )
+
     return {
         'canton': self.canton,
         'name': self.name,
-        'results': Archive(request.app.session()).latest_json(request)
+        'results': latest
     }